+++ swing/common-model/src/test/java/is/codion/swing/common/model/component/table/DefaultFilterTableModelTest.java
@@ -38,6 +38,7 @@ import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
 
+import static java.lang.System.lineSeparator;
 import static java.util.Arrays.asList;
 import static java.util.Collections.*;
 import static org.junit.jupiter.api.Assertions.*;
@@ -49,6 +50,7 @@ import static org.junit.jupiter.api.Assertions.*;
  */
 public final class DefaultFilterTableModelTest {
 
+	private static final String LINE_SEPARATOR = lineSeparator();
 	private static final TestRow A = new TestRow("a");
 	private static final TestRow B = new TestRow("b");
 	private static final TestRow C = new TestRow("c");
@@ -810,6 +812,56 @@ public final class DefaultFilterTableModelTest {
 		assertTrue(tableModel.items().contains(B));
 		assertTrue(tableModel.items().excluded().contains(B));
 		assertFalse(tableModel.items().contains(G));
+	}
+
+	@Test
+	void export() {
+		FilterTableModel<TestRow, Integer> table = createTestModel();
+		table.items().refresh();
+
+		String expected = "0" + LINE_SEPARATOR +
+						"a" + LINE_SEPARATOR +
+						"b" + LINE_SEPARATOR +
+						"c" + LINE_SEPARATOR +
+						"d" + LINE_SEPARATOR +
+						"e";
+		assertEquals(expected, table.export()
+						.delimiter('\t')
+						.get());
+
+		table.selection().indexes().set(asList(0, 1, 3));
+
+		String selected = "a" + LINE_SEPARATOR +
+						"b" + LINE_SEPARATOR +
+						"d";
+		assertEquals(selected, table.export()
+						.delimiter('\t')
+						.header(false)
+						.selected(true)
+						.get());
+
+		table.items().clear();
+
+		table.items().add(asList(
+						new TestRow("\nf\ng"),
+						new TestRow("g\r\nh\n"),
+						new TestRow("\ni\rj k")));
+
+		String result = "f g" + LINE_SEPARATOR +
+						"g h" + LINE_SEPARATOR +
+						"i j k";
+
+		assertEquals(result, table.export()
+						.header(false)
+						.get());
+
+		result = "f\ng" + LINE_SEPARATOR +
+						"g\r\nh" + LINE_SEPARATOR +
+						"i\rj k";
+		assertEquals(result, table.export()
+						.header(false)
+						.replaceNewline(null)
+						.get());
 	}
 
 	@Test